+++ packages/komodo_defi_rpc_methods/lib/src/common_structures/networks/lightning/channel/channels_index.dart
@@ -1,4 +1,4 @@
-export 'lightning_closed_channels_filter.dart';
-export 'lightning_open_channels_filter.dart';
 export 'lightning_channel_amount.dart';
 export 'lightning_channel_options.dart';
+export 'lightning_closed_channels_filter.dart';
+export 'lightning_open_channels_filter.dart';